% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtruncate.R
\name{dtruncate}
\alias{dtruncate}
\title{Compute density function of a truncated distribution}
\usage{
dtruncate(dist)
}
\arguments{
\item{dist}{character string indicating distribution name. It can be set as:
\itemize{
  \item beta Beta Distribution
  \item binom Binomial Distribution
  \item cauchy Cauchy Distribution
  \item chisq Chi-Square Distribution
  \item exp Exponential Distribution
  \item f F Distribution
  \item gamma Gamma Distribution
  \item geom Geometric Distribution
  \item hyper Hypergeometric Distribution
  \item logis Logistic Distribution
  \item lnorm Log Normal Distribution
  \item nbinom Negative Binomial Distribution
  \item norm Normal Distribution
  \item pois Poisson Distribution
  \item t Student t Distribution
  \item unif Uniform Distribution
  \item weibull Weibull Distribution
  \item wilcox Wilcoxon Rank Sum Statistic Distribution
  \item signrank Wilcoxon Signed Rank Statistic Distribution
}}
}
\value{
density function of the truncated specified distribution
}
\description{
Compute density function of a truncated distribution
}

